cleaned up compiler warnings

+++ spec/builtins/cfunc_spec.h
@@ -13,7 +13,7 @@ void testCfunc() {
     Xaddr my;
     stack_pop(r,XADDR_NOUN,&my);
     long *s = surface_for_xaddr(r,my);
-    spec_is_long_equal(*s,func);
+    spec_is_long_equal(*s,(long)func);
     op_invoke(r,my,RUN);
     char *peek_surface;
     stack_peek(r, CSTRING_NOUN, (void **)&peek_surface);
