+++ spec/receptor_util_spec.h
@@ -9,6 +9,6 @@ void testReceptorUtil(){
     Address to = {VM,0};
     Signal *s = signal_new(r,from,to,CSTRING_NOUN,"fish");
     spec_is_str_equal(&s->surface,"fish");
-    spec_is_equal(s->timestamp,t);
+    spec_is_equal((int)s->timestamp,(int)t);
     free(s);
 }
